% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart.R
\name{bar_chart}
\alias{bar_chart}
\title{Easy Bar Charts}
\usage{
bar_chart(
  data,
  x,
  y,
  facet,
  ...,
  bar_color = "#1F77B4",
  highlight = NULL,
  sort = TRUE,
  horizontal = TRUE,
  limit = NULL
)
}
\arguments{
\item{data}{Dataset to use for the bar chart}

\item{x}{The x variable}

\item{y}{numeric. If y is missing then it will be assigned the number of records in each group of y}

\item{facet}{A variable defining the faceting groups}

\item{...}{Additional arguments passed to aes()}

\item{bar_color}{character. The color of the bars.}

\item{highlight}{A value of \code{x} that should be highlighted in the plot}

\item{sort}{logical. Should the data be sorted before plotting?}

\item{horizontal}{logical. Should coord_flip() be added to the plot}

\item{limit}{integer. If a value for limit is provided only the first limit records will be displayed}
}
\description{
Easily create a bar chart
}
\examples{
data(biomedicalrevenue)
revenue2018 <- biomedicalrevenue[biomedicalrevenue$year == 2018, ]
revenue_roche <- biomedicalrevenue[biomedicalrevenue$company == "Roche", ]

## By default bar_chart() creates a horizontal and sorted plot
bar_chart(revenue2018, company, revenue)

## Create a vertical, non-sorted bar chart
bar_chart(revenue_roche, year, revenue, horizontal = FALSE, sort = FALSE)

## Limit the number of bars to the top 10
bar_chart(revenue2018, company, revenue, limit = 10)

## Change the bar color
bar_chart(revenue2018, company, revenue, bar_color = "purple")

## Highlight a single bar
bar_chart(revenue2018, company, revenue, limit = 10, highlight = "Roche")

## Use facets to show the top 10 companies over the years
bar_chart(biomedicalrevenue, company, revenue, facet = year, limit = 10)

}
\author{
Thomas Neitmann
}
