% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomedicalrevenue}
\alias{biomedicalrevenue}
\title{Top Biomedical Companies Revenues.}
\format{A data frame with 224 rows and 3 variables:
\describe{
  \item{company}{Name of the company}
  \item{year}{Fiscal year}
  \item{revenue}{Revenue in billion USD}
}}
\source{
\url{https://en.wikipedia.org/wiki/List_of_largest_biomedical_companies_by_revenue}
}
\usage{
biomedicalrevenue
}
\description{
A dataset containing the annual revenues of top biomedical companies from 2011 to 2018.
}
\keyword{datasets}
