% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diverging_bar_chart.R
\name{diverging_bar_chart}
\alias{diverging_bar_chart}
\title{Diverging Bar Chart}
\usage{
diverging_bar_chart(
  data,
  x,
  y,
  bar_colors = c("#1F77B4", "#FF7F0E"),
  text_color = "black",
  text_size = 10
)
}
\arguments{
\item{data}{Dataset to use for the chart}

\item{x}{The x variable}

\item{y}{The y variable}

\item{bar_colors}{A vector of length 2 containing the colors for the positive
and negative bars}

\item{text_color}{The color for the bar annotations}

\item{text_size}{The size of the bar annotation text in pt}
}
\description{
Easily create a diverging bar chart
}
\examples{
if (requireNamespace("tidyr")) {
  `\%>\%` <- magrittr::`\%>\%`
  biomedicalrevenue \%>\%
  dplyr::filter(year > 2016) \%>\%
  tidyr::pivot_wider(
    values_from = revenue,
    names_from = year,
    names_prefix = "revenue_"
  ) \%>\%
  dplyr::mutate(diff = revenue_2018 - revenue_2017) \%>\%
  diverging_bar_chart(company, diff)
}

}
