% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diverging_lollipop_chart.R
\name{diverging_lollipop_chart}
\alias{diverging_lollipop_chart}
\title{Diverging Lollipop Chart}
\usage{
diverging_lollipop_chart(
  data,
  x,
  y,
  lollipop_colors = c("#1F77B4", "#FF7F0E"),
  line_size = 0.75,
  point_size = 3,
  text_color = "black",
  text_size = 10
)
}
\arguments{
\item{data}{Dataset to use for the chart}

\item{x}{The x variable}

\item{y}{The y variable}

\item{lollipop_colors}{A vector of length 2 containing the colors for the positive
and negative lollipops}

\item{line_size}{numeric. Size of the lollipop 'stick'}

\item{point_size}{numeric. Size of the lollipop 'head'}

\item{text_color}{The color for the lollipop annotations}

\item{text_size}{The size of the lollipop annotation text in pt}
}
\description{
Easily create a diverging lollipop chart
}
\examples{
`\%>\%` <- magrittr::`\%>\%`
if (requireNamespace("tidyr")) {
  data(biomedicalrevenue)
  biomedicalrevenue \%>\%
  dplyr::filter(year > 2016) \%>\%
  tidyr::pivot_wider(
    values_from = revenue,
    names_from = year,
    names_prefix = "revenue_"
  ) \%>\%
  dplyr::mutate(diff = revenue_2018 - revenue_2017) \%>\%
  diverging_lollipop_chart(company, diff)
}

data(mtcars)
mtcars \%>\%
  dplyr::mutate(model = row.names(.), hpz = scale(hp)) \%>\%
  diverging_lollipop_chart(model, hpz)

}
