% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumbbell_chart.R
\name{dumbbell_chart}
\alias{dumbbell_chart}
\title{Dumbbell Chart}
\usage{
dumbbell_chart(
  data,
  x,
  y1,
  y2,
  line_size = 1.5,
  line_color = "gray",
  point_size = 4,
  point_colors = c("#1F77B4", "#FF7F0E"),
  sort = TRUE,
  horizontal = TRUE,
  limit = NULL,
  legend = TRUE
)
}
\arguments{
\item{data}{Dataset to use for plot}

\item{x}{Variable representing the categories to plot}

\item{y1}{Variable representing the dumbbell end}

\item{y2}{Variable representing the dumbbell start}

\item{line_size}{\code{numeric}. Line width}

\item{line_color}{\code{character}. Line color}

\item{point_size}{\code{numeric}. Point size}

\item{point_colors}{\code{numeric}. Point color}

\item{sort}{\code{logical}. Should the data be sorted before plotting?}

\item{horizontal}{\code{logical}. Should the plot be displayed horizontally?}

\item{limit}{\code{integer}. f a value for limit is provided only the first limit records will be displayed}

\item{legend}{\code{logical}. Should a legend be displayed?}
}
\description{
Easily create a dumbbell chart
}
\examples{
if (requireNamespace("tidyr")) {
  library(magrittr)
  data(biomedicalrevenue)

  biomedicalrevenue \%>\%
    dplyr::filter(year \%in\% c(2011, 2018)) \%>\%
    tidyr::pivot_wider(
      values_from = revenue,
      names_from = year,
      names_prefix = "revenue_"
    ) \%>\%
    dplyr::filter(revenue_2011 > 40, revenue_2018 > 40) \%>\%
    dumbbell_chart(company, revenue_2011, revenue_2018)
}

}
