% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggcharts_get_default_color}
\alias{ggcharts_get_default_color}
\title{Get the Default Color for a ggcharts Theme}
\usage{
ggcharts_get_default_color(theme)
}
\arguments{
\item{theme}{\code{character}. The name of a \code{ggcharts} theme.}
}
\value{
The default color for the given theme as a \code{character}
}
\description{
Retrieve the color used by default for a given \code{ggcharts} theme
}
\examples{
ggcharts_get_default_color("theme_hermit")
ggcharts_get_default_color("theme_ng")

}
\author{
Thomas Neitmann
}
