% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipop_chart.R
\name{lollipop_chart}
\alias{lollipop_chart}
\title{Lollipop Chart}
\usage{
lollipop_chart(
  data,
  x,
  y,
  ...,
  line_size = 0.75,
  line_color = "#1F77B4",
  point_size = 4,
  point_color = line_color,
  sort = TRUE,
  horizontal = TRUE,
  limit = NULL
)
}
\description{
Easily create a lollipop chart using ggplot2
}
