% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popch}
\alias{popch}
\title{Population Statistics of Switzerland}
\format{
A data frame with 42 rows and 3 variables:
\describe{
  \item{age}{Five-year age group}
  \item{sex}{Sex}
  \item{pop}{Population}
}
}
\source{
US Census International Data Base
}
\usage{
popch
}
\description{
Swiss population in 2020 by five-year age groups
}
\keyword{datasets}
