% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popeurope}
\alias{popeurope}
\title{European Population}
\format{
A data frame with 30 rows and 3 variables:
\describe{
  \item{country}{Name of the country}
  \item{pop1952}{Population in 1952}
  \item{pop2007}{Population in 2007}
}
}
\source{
\url{http://www.gapminder.org/data/}
}
\usage{
popeurope
}
\description{
Population of European countries in 1952 and 2007
}
\keyword{datasets}
