% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_plot.R
\name{post_process_plot}
\alias{post_process_plot}
\title{importFrom ggplot2 coord_flip scale_fill_manual facet_wrap}
\usage{
post_process_plot(plot, horizontal, facet, fill)
}
\description{
importFrom ggplot2 coord_flip scale_fill_manual facet_wrap
}
