% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid_chart.R
\name{pyramid_chart}
\alias{pyramid_chart}
\title{Pyramid Chart}
\usage{
pyramid_chart(
  data,
  x,
  y,
  group,
  bar_colors = c("#1F77B4", "#FF7F0E"),
  sort = "no",
  title = NULL,
  xlab = NULL
)
}
\arguments{
\item{data}{Dataset to use for the pyramid chart}

\item{x}{\code{character} or \code{factor} column of \code{data}}

\item{y}{\code{numeric} column of \code{data}}

\item{group}{\code{character} or \code{factor} column of \code{data}}

\item{bar_colors}{\code{character} vector of length 2 containing colors}

\item{sort}{\code{character}. Should the bars be sorted? By default \code{"no"}.}

\item{title}{\code{character}. Deprecated. Plot title. By default no title is displayed.}

\item{xlab}{\code{character}. Deprecated. X axis label.}
}
\value{
An object of class \code{PyramidChart}
}
\description{
Easily create a pyramid chart
}
\details{
Arguments \code{xlab} and \code{title} are soft-deprecated, i.e. you can
still use them but a warning will be thrown. Instead, use \code{labs}. See
below for an example.
}
\examples{
library(magrittr)
data(mtcars)
cars <- mtcars \%>\%
  dplyr::count(cyl, am) \%>\%
  dplyr::mutate(am = ifelse(am == 0, "Manual", "Automatic"))

pyramid_chart(cars, cyl, n, am)

pyramid_chart(cars, cyl, n, am) +
  labs(title = "Pyramid Chart", x = "Number of Cars")

}
\author{
Thomas Neitmann
}
