% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PyramidChart.R
\docType{class}
\name{PyramidChart-class}
\alias{PyramidChart-class}
\alias{PyramidChart}
\alias{+,PyramidChart,theme-method}
\alias{+,PyramidChart,ScaleContinuousPosition-method}
\alias{+,PyramidChart,labels-method}
\alias{show,PyramidChart-method}
\title{An S4 Class Representing a Pyramid Chart}
\usage{
\S4method{+}{PyramidChart,theme}(e1, e2)

\S4method{+}{PyramidChart,ScaleContinuousPosition}(e1, e2)

\S4method{+}{PyramidChart,labels}(e1, e2)

\S4method{show}{PyramidChart}(object)
}
\arguments{
\item{e1}{An object of class \code{PyramidChart}}

\item{e2}{An object of class \code{labels}}

\item{object}{An object of class \code{PyramidChart}}
}
\description{
An S4 Class Representing a Pyramid Chart

Change PyramidChart theme

Change PyramidChart scale

Change PyramidChart labels

Print PyramidChart
}
\section{Slots}{

\describe{
\item{\code{plot}}{The pyramid chart created with \code{pyramic_chart}}
}}

