% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_chart.R
\name{scatter_chart}
\alias{scatter_chart}
\title{Scatter Chart}
\usage{
scatter_chart(
  data,
  x,
  y,
  group,
  point_color = "auto",
  point_size = 2,
  jitter = FALSE
)
}
\arguments{
\item{data}{Dataset to use for the bar chart}

\item{x}{\code{numeric} column of \code{data}}

\item{y}{\code{numeric} column of \code{data}}

\item{group}{\code{character} or \code{factor} column of \code{data}}

\item{point_color}{\code{character}. Point color}

\item{point_size}{\code{numeric}. Point diameter in mm}

\item{jitter}{\code{logical}. Should jitter be applied to the points?}
}
\value{
An object of class \code{ggplot}
}
\description{
Easily create a scatter chart
}
\examples{
data(iris)

scatter_chart(iris, Sepal.Width, Sepal.Length)

scatter_chart(iris, Sepal.Width, Sepal.Length,
              jitter = TRUE, point_size = 3)

scatter_chart(iris, Sepal.Width, Sepal.Length,
              point_color = "darkgreen")

scatter_chart(iris, Sepal.Width, Sepal.Length,
              group = Species)

}
\author{
Thomas Neitmann
}
