% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_ggcharts}
\alias{theme_ggcharts}
\alias{theme_hermit}
\title{Themes}
\usage{
theme_ggcharts(
  base_size = 14,
  base_family = "",
  axis_line = NA,
  grid_line = NA
)

theme_hermit(base_size = 14, base_family = "", axis_line = NA, grid_line = NA)
}
\arguments{
\item{base_size}{\code{numeric}. Base font size in pt.}

\item{base_family}{\code{character}. Base font family.}

\item{axis_line}{\code{character}. Where to draw an axis line.}

\item{grid_line}{\code{character}. Where to draw grid lines.}
}
\description{
ggplot2 themes
}
\details{
\code{theme_ggcharts} is the default theme used when creating any plot with
\code{ggcharts}.
Accepted values for \code{axis_line} and \code{grid_line} are
\code{c(NA, "X", "Y", "XY")}. When set to \code{NA}, the default, nothing is
drawn.
}
\examples{
library(ggplot2)
library(dplyr)

scatter1 <- ggplot(mtcars, aes(hp, mpg)) +
  geom_point(color = "steelblue")

scatter1 + theme_ggcharts()

scatter1 + theme_ggcharts(grid_line = "XY")

scatter1 + theme_ggcharts(axis_line = "XY")

scatter2 <- ggplot(mtcars, aes(hp, mpg)) +
  geom_point(color = "yellow")

scatter2 + theme_hermit()

scatter2 + theme_hermit(grid_line = "XY")

scatter2 + theme_hermit(axis_line = "XY")

bar_chart(ggplot2::diamonds, cut, bar_color = "orange") +
  theme_hermit(axis_line = "Y", grid_line = "Y")

column_chart(ggplot2::diamonds, cut, bar_color = "steelblue") +
  theme_ggcharts(axis_line = "X", grid_line = "X")

ggcharts::biomedicalrevenue \%>\%
  filter(company == "Roche") \%>\%
  ggplot(aes(year, revenue)) +
  geom_line(color = "yellow", size = 1) +
  scale_y_continuous(expand = expand_scale(c(0, .05))) +
  theme_hermit(grid_line = "X", axis_line = "X")

}
\author{
Thomas Neitmann
}
