% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_coffee}
\alias{theme_coffee}
\title{Theme Coffee}
\usage{
theme_coffee(
  base_size = 14,
  base_family = "Cooper Hewitt",
  header_family = "Cooper Hewitt",
  axis = "",
  ticks = "",
  grid = ""
)
}
\arguments{
\item{base_size}{\code{numeric}. Base font size in pt}

\item{base_family}{\code{character}. Base font family}

\item{header_family}{\code{character}. Font family for title and subtitle}

\item{axis}{\code{character}. Where to draw an axis line}

\item{ticks}{\code{character}. Where to draw axis ticks}

\item{grid}{\code{character}. Where to draw grid lines}
}
\value{
An object of class \code{theme}
}
\description{
A theme inspired by coffee
}
\examples{
library(ggplot2)
library(dplyr)

scatter <- ggplot(mtcars, aes(hp, mpg)) +
  geom_point(color = "#F4C95D")

scatter + theme_coffee()

scatter + theme_coffee(grid = "XY")

scatter + theme_coffee(axis = "xy", ticks = "xy")

bar_chart(ggplot2::diamonds, cut, bar_color = "#F4C95D") +
  theme_coffee(axis = "y", grid = "Y")

column_chart(ggplot2::diamonds, cut, bar_color = "#F4C95D") +
  theme_coffee(axis = "x", grid = "X")

ggcharts::biomedicalrevenue \%>\%
  filter(company == "Roche") \%>\%
  ggplot(aes(year, revenue)) +
  geom_line(color = "#F4C95D", size = 1) +
  scale_y_continuous(expand = expand_scale(c(0, .05))) +
  theme_coffee(grid = "X", axis = "x", ticks = "x")

}
\seealso{
For more details see the 'theme' vignette:
\code{vignette("theme", package = "ggcharts")}
}
\author{
Thomas Neitmann
}
