% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipop_chart.R
\name{lollipop_chart}
\alias{lollipop_chart}
\title{Lollipop Chart}
\usage{
lollipop_chart(
  data,
  x,
  y,
  facet,
  ...,
  line_size = 0.75,
  line_color = "#1F77B4",
  point_size = 4,
  point_color = line_color,
  highlight = NULL,
  sort = TRUE,
  horizontal = TRUE,
  limit = NULL
)
}
\arguments{
\item{data}{Dataset to use for the bar chart}

\item{x}{The x variable}

\item{y}{numeric. If y is missing then it will be assigned the number of records in each group of y}

\item{facet}{A variable defining the faceting groups}

\item{...}{Additional arguments passed to aes()}

\item{line_size}{character. Size of the lollipop 'stick'}

\item{line_color}{character. Color of the lollipop 'stick'}

\item{point_size}{character. Size of the lollipop 'head'}

\item{point_color}{character. Color of the lollipop 'head'}

\item{highlight}{A value of \code{x} that should be highlighted in the plot}

\item{sort}{logical. Should the data be sorted before plotting?}

\item{horizontal}{logical. Should coord_flip() be added to the plot}

\item{limit}{integer. If a value for limit is provided only the first limit records will be displayed}
}
\description{
Easily create a lollipop chart using ggplot2
}
\examples{
library(dplyr)
data(biomedicalrevenue)
biomedicalrevenue \%>\%
  filter(year == 2015) \%>\%
  bar_chart(company, year, limit = 10)

}
\author{
Thomas Neitmann
}
