% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_ggcharts}
\alias{theme_ggcharts}
\alias{theme_hermit}
\title{Themes}
\usage{
theme_ggcharts(
  base_size = 14,
  base_family = "",
  axis_line = NA,
  grid_line = NA
)

theme_hermit(base_size = 14, base_family = "", axis_line = NA, grid_line = NA)
}
\arguments{
\item{base_size}{\code{numeric}. Base font size in pt.}

\item{base_family}{\code{character}. Base font family.}

\item{axis_line}{\code{character}. Where to draw an axis line.}

\item{grid_line}{\code{character}. Where to draw grid lines.}
}
\description{
ggplot2 themes
}
\details{
\code{theme_ggcharts} is the default theme used when creating any plot with
\code{ggcharts}.
Accepted values for \code{axis_line} and \code{grid_line} are
\code{c(NA, "X", "Y", "XY")}. When set to \code{NA}, the default, nothing is
drawn.
}
\author{
Thomas Neitmann
}
